Fix/comment

+++ src/main/java/com/example/sendowl/api/service/CommentService.java
@@ -3,15 +3,13 @@ package com.example.sendowl.api.service;
 import com.example.sendowl.domain.board.exception.enums.BoardErrorCode;
 import com.example.sendowl.domain.board.entity.Board;
 import com.example.sendowl.domain.comment.entity.Comment;
+import com.example.sendowl.domain.comment.exception.CommentNoPermission;
 import com.example.sendowl.domain.comment.exception.CommentNotFoundException;
 import com.example.sendowl.domain.comment.exception.enums.CommentErrorCode;
 import com.example.sendowl.domain.user.entity.User;
 import com.example.sendowl.domain.board.exception.BoardNotFoundException;
 import com.example.sendowl.domain.board.repository.BoardRepository;
 import com.example.sendowl.domain.comment.repository.CommentRepository;
-import com.example.sendowl.domain.user.exception.UserException.*;
-import com.example.sendowl.domain.user.exception.enums.UserErrorCode;
-import com.example.sendowl.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,17 +25,13 @@ import java.util.List;
 public class CommentService {
 
     private final CommentRepository commentRepository;
-    private final UserRepository userRepository;
     private final BoardRepository boardRepository;
 
     @Transactional
-    public CommentRes insertComment(CommentReq vo) {
+    public CommentRes insertComment(CommentReq vo, User user) {
 
         Board board = boardRepository.findById(vo.getBoardId())
                 .orElseThrow(()->new BoardNotFoundException(BoardErrorCode.NOT_FOUND));
-
-        User user = userRepository.findByEmail(vo.getEmail())
-                .orElseThrow(()->new UserNotFoundException(UserErrorCode.NOT_FOUND));
 
         Comment comment;
 
@@ -57,6 +51,7 @@ public class CommentService {
                     .board(board)
                     .user(user)
                     .content(vo.getContent())
+                    .parent(null)
                     .depth(0L)
                     .build();
         }
@@ -84,27 +79,34 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentRes updateComment(UpdateReq crq) {
+    public CommentRes updateComment(UpdateReq crq, User user) {
 
         Comment updComment = commentRepository.findById(crq.getCommentId()).orElseThrow(
                 () -> new CommentNotFoundException(CommentErrorCode.NOT_FOUND));
 
-        updComment.updateContent(crq.getContent());
-
-        commentRepository.save(updComment);
-
-        CommentRes commentRes = new CommentRes(updComment);
-
-        return commentRes;
+        // 글쓴 유저가 맞는지 확인하기
+        if(updComment.getUser().getId() == user.getId()){
+            updComment.updateContent(crq.getContent());
+            commentRepository.save(updComment);
+            CommentRes commentRes = new CommentRes(updComment);
+            return commentRes;
+        }
+        else{
+            throw new CommentNoPermission(CommentErrorCode.NO_PERMISSON);
+        }
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, User user) {
         Comment delComment = commentRepository.findById(commentId).orElseThrow(
                 () -> new CommentNotFoundException(CommentErrorCode.NOT_FOUND));
 
-        delComment.delete();
-
-        commentRepository.save(delComment);
+        if(delComment.getUser().getId() == user.getId()){
+            delComment.delete();
+            commentRepository.save(delComment);
+        }
+        else{
+            throw new CommentNoPermission(CommentErrorCode.NO_PERMISSON);
+        }
     }
 }
